# api/userProgram/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a swag test http server.",
        "title": "USER WEB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7001",
    "basePath": "/api/v1",
    "paths": {
...
        "/home/banner/getList": {
            "get": {
                "description": "获取首页Banner列表 (TODO：返回参数)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取首页Banner列表",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appConfigService.GetBannerListModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/home/config/list": {
            "get": {
                "description": "app 配置比如首页分类相关等配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取app配置列表",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appConfigService.GetAppConfigListModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.GetAppConfigListResponseModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/home/config/product/list": {
            "get": {
                "description": "获取首页配置产品信息列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取首页配置产品信息列表",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appConfigService.GetHomeConfigProductInfoModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/productService.ProductSmallInfoModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/home/productConfig/getList": {
            "get": {
                "description": "获取首页产品配置列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取首页产品配置列表",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appConfigService.GetHomeProductConfigListModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.HomeProductConfig"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/hot/search/get": {
            "get": {
                "description": "描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "简单描述",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/controller.GetSearchHotListModel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
        "/wechat/getHomeData": {
            "get": {
                "description": "待确定",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取首页 hometba 配置",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.AppConfig"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "appConfigService.GetAppConfigListModel": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "appConfigService.GetBannerListModel": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "appConfigService.GetHomeConfigProductInfoModel": {
            "type": "object",
            "properties": {
                "appConfigId": {
                    "type": "integer"
                }
            }
        },
        "appConfigService.GetHomeProductConfigListModel": {
            "type": "object",
            "properties": {
                "appConfigId": {
                    "type": "integer"
                }
            }
        },
...
        "controller.GetAppConfigListResponseModel": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "textList": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "controller.GetSearchHotListModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.AppConfig": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "configType": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "linkType": {
                    "type": "integer"
                },
                "linkUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "productSmallInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/productService.ProductSmallInfoModel"
                    }
                },
                "resourceId": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.HomeProductConfig": {
            "type": "object",
            "properties": {
                "appConfigId": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                }
            }
        },
...
        "productService.ProductSmallInfoModel": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "object"
                },
                "brandId": {
                    "type": "integer"
                },
                "currentPrice": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "guid": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "keywords": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "maxPrice": {
                    "type": "number"
                },
                "merId": {
                    "type": "integer"
                },
                "merName": {
                    "type": "string"
                },
                "minPrice": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "originalPrice": {
                    "type": "number"
                },
                "productType": {
                    "type": "integer"
                },
                "sales": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "typeId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}